requirements.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.15
psycopg2-binary==2.9.11
pillow==12.1.1
PyJWT==2.12.1
PyYAML==6.0.3
sqlparse==0.5.5
asgiref==3.11.1
uritemplate==4.2.0
inflection==0.5.1
packaging==26.0
pytz==2026.1.post1